Wrap the sample SRO output to make it a full document
Seems that the basic format doesn't include <html>, <body> etc
tags. Not yet sure whether that is really how it's going to work. I'm
adding them manually here so we can get on with working out the
styling -- we can always change the import code later to cope with
their absence.